feat: endpoint user BIASALAH

+++ database/migration/000001_create_users_table.up.sql
@@ -1,5 +1,5 @@
 create table if not exists users (
-   user_id bigint primary key,
+   user_id serial primary key,
    email varchar unique not null,
    name varchar not null
 );
